refactor: move assert_downloaded to test_helper.rb so other tests can use it.
